Move explorer to js module
