+++ docs/2025/index.html
@@ -171,6 +171,6 @@ style="margin-top: 1rem; font-style: italic;">
 <table id="query-results" class="data-table">
 </table>
 </div>
-<script src="main.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
